+++ contracts/metatx/ManagedIdentity.sol
@@ -14,7 +14,6 @@ abstract contract ManagedIdentity {
     }
 
     function _msgData() internal view virtual returns (bytes memory) {
-        this; // silence state mutability warning without generating bytecode - see https://github.com/ethereum/solidity/issues/2691
         return msg.data;
     }
 }